docs-gen/requirements.txt
pyyaml==6.0.2
markdown==3.6
openai==1.43.0
anthropic==0.34.0
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
